requirements.txt
# ===== Dépendances principales =====
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# ===== Traitement d'images =====
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# ===== Data science =====
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0

# ===== Logging & monitoring =====
wandb>=0.15.0
tensorboard>=2.14.0

# ===== Configuration =====
pyyaml>=6.0
omegaconf>=2.3.0

# ===== Versioning données =====
dvc>=3.0.0

# ===== Utilitaires =====
tqdm>=4.65.0
colorama>=0.4.6
einops>=0.6.0

# ===== Tests =====
pytest>=7.4.0
pytest-cov>=4.1.0

# ===== Notebooks =====
jupyter>=1.0.0
ipywidgets>=8.0.0
